#136: Move to Qt 6 (#185)
# Description
This commit makes a series of changes to move the app to compile with Qt
6.8, as Qt 5.15 will [no longer be supported after 26 May
2025](https://www.qt.io/blog/extended-security-maintenance-for-qt-5.15-begins-may-2025).
It also bumps the C++ version to C++20. Deprecated APIs are also removed
(codecvt in the C++ standard library, `NSColor : colorSpaceName` and
`NSAppearance : currentAppearance` on macOS).

Closes #136.

## Type of change
- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] This change requires a documentation update

# How Has This Been Tested?

All tests in the test suite pass, and this has been manually tested on
macOS, Linux, and Windows

# Checklist:

- [x] My code follows the style guidelines of this project (`black` for
Python
  code, `.clang-format` in the `src/jyut-dict` directory for C++)
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have translated my user-facing strings to all
currently-supported languages
- [x] I have made corresponding changes to the documentation