#41: Add fuzzy Jyutping searching (#186)
# Description

Though Jyutping is the most commonly used standard for Cantonese
romanization, many speakers and learners have little exposure to it. As
a result, when attempting to search, most will generally use an ad-hoc
romanization, informed by the sounds of English. This commit adds
functionality to enable searching in non-standard romanization.

Changes are as follows:
* REGEXP is used for the SQL query instead of GLOB.
* Search terms undergo one round of "corrections" before syllable
segmentation.
* Segmented syllables are then processed further to account for "lazy
pronunciation" sounds.

The results are that search is now more forgiving, at the cost of the
query being more expensive (as REGEXP does not use database indices) and
matching entries that were not what the user was searching for. Note
that performance is significantly worse on Windows (haven't figured out
why), and will likely be disabled by default or chosen by the user in a
future commit.

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# How Has This Been Tested?

Multiple unit tests have been added, and the feature has been manually
tested on all three operating systems.

# Checklist:

- [x] My code follows the style guidelines of this project (`black` for
Python
  code, `.clang-format` in the `src/jyut-dict` directory for C++)
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have translated my user-facing strings to all
currently-supported languages
- [x] I have made corresponding changes to the documentation