#41: Add UI for fuzzy Jyutping (#187)
# Description

This commit adds a UI for the new fuzzy Jyutping feature. It adds a new
tab in settings to enable/disable fuzzy Jyutping and dangerous fuzzy
Jyutping substitutions. It also adds a pop-up on first launch to enable
a user to choose whether to enable this feature.

Closes #41.

## Type of change

- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)

# How Has This Been Tested?

This feature has been tested extensively on macOS, Linux, and Windows.

# Checklist:

- [x] My code follows the style guidelines of this project (`black` for
Python
  code, `.clang-format` in the `src/jyut-dict` directory for C++)
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have translated my user-facing strings to all
currently-supported languages
- [x] I have made corresponding changes to the documentation