#189: Enable searching using dictation (#192)
# Description

This commit integrates with Speech Recognition APIs on macOS to enable
searching using dictation. On Windows, it activates the keyboard
corresponding to the target language, and then sends the Win+H key combo
to start dictation. On Linux, this feature is not supported.

Closes #189.

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# How Has This Been Tested?

Tested on macOS 15.4, Windows 11 23H2, Fedora 39.

# Checklist:

- [x] My code follows the style guidelines of this project (`black` for
Python
  code, `.clang-format` in the `src/jyut-dict` directory for C++)
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have translated my user-facing strings to all
currently-supported languages
- [x] I have made corresponding changes to the documentation